applying requested changes and update docs
